fix: Notification observing on app hidden: moved persistent connection back to viewModel
