+++ app/src/main/kotlin/com/wire/android/ui/WireActivity.kt
@@ -111,6 +111,7 @@ class WireActivity : AppCompatActivity() {
         proximitySensorManager.initialize()
         lifecycle.addObserver(currentScreenManager)
 
+        viewModel.observePersistentConnectionStatus()
         val startDestination = viewModel.startNavigationRoute()
         setComposableContent(startDestination) {
             handleDeepLink(intent, savedInstanceState)
