+++ app/src/test/kotlin/com/wire/android/GlobalObserversManagerTest.kt
@@ -47,49 +47,6 @@ class GlobalObserversManagerTest {
     }
 
     @Test
-    fun `given valid accounts, all with persistent socket disabled, then stop socket service`() {
-        val statuses = listOf(
-            PersistentWebSocketStatus(TestUser.SELF_USER.id, false),
-            PersistentWebSocketStatus(TestUser.USER_ID.copy(value = "something else"), false)
-        )
-        val accounts = listOf(
-            TestUser.SELF_USER,
-            TestUser.SELF_USER.copy(id = TestUser.USER_ID.copy(value = "something else"))
-        )
-        val (arrangement, manager) = Arrangement()
-            .withValidAccounts(accounts.map { it to null })
-            .withPersistentWebSocketConnectionStatuses(statuses)
-            .arrange()
-
-        manager.observe()
-
-        coVerify(exactly = 0) { arrangement.servicesManager.startPersistentWebSocketService() }
-        coVerify(exactly = 1) { arrangement.servicesManager.stopPersistentWebSocketService() }
-    }
-
-    @Test
-    fun `given valid accounts, at least one with persistent socket enabled, and socket service not running, then start service`() {
-        val statuses = listOf(
-            PersistentWebSocketStatus(TestUser.SELF_USER.id, false),
-            PersistentWebSocketStatus(TestUser.USER_ID.copy(value = "something else"), true)
-        )
-        val accounts = listOf(
-            TestUser.SELF_USER,
-            TestUser.SELF_USER.copy(id = TestUser.USER_ID.copy(value = "something else"))
-        )
-        val (arrangement, manager) = Arrangement()
-            .withValidAccounts(accounts.map { it to null })
-            .withPersistentWebSocketConnectionStatuses(statuses)
-            .withIsPersistentWebSocketServiceRunning(false)
-            .arrange()
-
-        manager.observe()
-
-        coVerify(exactly = 1) { arrangement.servicesManager.startPersistentWebSocketService() }
-        coVerify(exactly = 0) { arrangement.servicesManager.stopPersistentWebSocketService() }
-    }
-
-    @Test
     fun `given valid accounts, at least one with persistent socket enabled, and socket service running, then do not start service again`() {
         val statuses = listOf(
             PersistentWebSocketStatus(TestUser.SELF_USER.id, false),
@@ -102,13 +59,10 @@ class GlobalObserversManagerTest {
         val (arrangement, manager) = Arrangement()
             .withValidAccounts(accounts.map { it to null })
             .withPersistentWebSocketConnectionStatuses(statuses)
-            .withIsPersistentWebSocketServiceRunning(true)
             .arrange()
 
         manager.observe()
 
-        coVerify(exactly = 0) { arrangement.servicesManager.startPersistentWebSocketService() }
-        coVerify(exactly = 0) { arrangement.servicesManager.stopPersistentWebSocketService() }
         coVerify(exactly = 1) {
             arrangement.notificationManager.observeNotificationsAndCallsWhileRunning(
                 listOf(TestUser.SELF_USER.id),
@@ -132,17 +86,13 @@ class GlobalObserversManagerTest {
         @MockK
         lateinit var navigationManager: NavigationManager
 
-        @MockK
-        lateinit var servicesManager: ServicesManager
-
         private val manager by lazy {
             GlobalObserversManager(
                 dispatcherProvider = TestDispatcherProvider(),
                 coreLogic = coreLogic,
                 notificationChannelsManager = notificationChannelsManager,
                 notificationManager = notificationManager,
-                navigationManager = navigationManager,
-                servicesManager = servicesManager
+                navigationManager = navigationManager
             )
         }
 
@@ -157,10 +107,6 @@ class GlobalObserversManagerTest {
 
         fun withValidAccounts(list: List<Pair<SelfUser, Team?>>): Arrangement = apply {
             coEvery { coreLogic.getGlobalScope().observeValidAccounts() } returns flowOf(list)
-        }
-
-        fun withIsPersistentWebSocketServiceRunning(isRunning: Boolean): Arrangement = apply {
-            every { servicesManager.isPersistentWebSocketServiceRunning() } returns isRunning
         }
 
         fun withPersistentWebSocketConnectionStatuses(list: List<PersistentWebSocketStatus>): Arrangement = apply {
